requirements.txt

fastapi==0.115.5
uvicorn==0.32.1
langchain==0.3.13
langchain-core==0.3.28
langchain-google-genai==2.0.8
pypdf==5.1.0
python-multipart==0.0.20
google-generativeai==0.8.3
pydantic==2.10.3
jinja2==3.1.4
pandas==2.2.3
faiss-cpu==1.9.0
numpy==2.2.1
langchain-community==0.3.13
langchain-chroma==0.1.4
langgraph==0.2.60
